requirements.txt
gmqtt>=0.6.0
asyncpg>=0.29.0
pyyaml>=6.0
aiocoap>=0.4.7
DTLSSocket>=0.1.16
cryptography>=42.0.0
requests>=2.31.0
